+++ CHANGELOG.md
@@ -18,6 +18,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed bugs
 
+* `jj bookmark forget` no longer prints `Forgot N local bookmarks.` when no
+  local bookmarks were actually forgotten (e.g. when only an untracked remote
+  bookmark matched). [#9181](https://github.com/jj-vcs/jj/issues/9181).
+
 ## [0.41.0] - 2026-05-06
 
 ### Release highlights